+++ vitest.config.ts
@@ -1,0 +1,17 @@
+import { defineConfig } from 'vitest/config'
+
+export default defineConfig({
+  test: {
+    include: ['test/**/*.test.ts'],
+    environment: 'node',
+    passWithNoTests: true,
+  },
+  resolve: {
+    alias: [
+      {
+        find: /^#exjs-controllers\/(.*)$/,
+        replacement: `${process.cwd()}/$1.ts`,
+      },
+    ],
+  },
+})